added nfs

+++ libs/libsos/src/sos_file.c
@@ -12,7 +12,7 @@ int sos_stat(const char *path, sos_stat_t *buf) {
 }
 
 int sos_getdirent(int pos, char *name, size_t nbyte) {
-	return SYSCALL_ARG3(SOS_SYSCALL_GETDIRENT, pos, name, nbyte);
+	return SYSCALL_ARG3(SOS_SYSCALL_GETDIRENT, pos, (int)name, nbyte);
 }
 
 int sos_sys_close(int file) { return SYSCALL_ARG1(SOS_SYSCALL_CLOSE, file); }
